Rename macro with a mmtk prefix. Move macro impl to a separate module.
