+++ macros/trace_object/src/derive_impl.rs
@@ -1,0 +1,130 @@
+use quote::quote;
+use syn::{Field, TypeGenerics};
+use proc_macro2::TokenStream as TokenStream2;
+
+use crate::util;
+
+pub fn generate_trace_object<'a>(
+    space_fields: &[&'a Field],
+    parent_field: &Option<&'a Field>,
+    ty_generics: &TypeGenerics,
+) -> TokenStream2 {
+    // Generate a check with early return for each space
+    let space_field_handler = space_fields.iter().map(|f| {
+        let f_ident = f.ident.as_ref().unwrap();
+        let ref f_ty = f.ty;
+
+        // Figure out copy
+        let trace_attr = util::get_field_attribute(f, "trace").unwrap();
+        let copy = if !trace_attr.tokens.is_empty() {
+            use syn::Token;
+            use syn::NestedMeta;
+            use syn::punctuated::Punctuated;
+
+            let args = trace_attr.parse_args_with(Punctuated::<NestedMeta, Token![,]>::parse_terminated).unwrap();
+            // CopySemantics::X is a path.
+            if let Some(NestedMeta::Meta(syn::Meta::Path(p))) = args.first() {
+                quote!{ Some(#p) }
+            } else {
+                quote!{ None }
+            }
+        } else {
+            quote!{ None }
+        };
+
+        quote! {
+            if self.#f_ident.in_space(__mmtk_objref) {
+                return <#f_ty as PolicyTraceObject #ty_generics>::trace_object::<T, KIND>(&self.#f_ident, __mmtk_trace, __mmtk_objref, #copy, __mmtk_worker);
+            }
+        }
+    });
+
+    // Generate a fallback to the parent plan
+    let parent_field_delegator = if let Some(f) = parent_field {
+        let f_ident = f.ident.as_ref().unwrap();
+        let ref f_ty = f.ty;
+        quote! {
+            <#f_ty as PlanTraceObject #ty_generics>::trace_object::<T, KIND>(&self.#f_ident, __mmtk_trace, __mmtk_objref, __mmtk_worker)
+        }
+    } else {
+        quote! {
+            panic!("No more spaces to try")
+        }
+    };
+
+    quote! {
+        fn trace_object<T: crate::plan::TransitiveClosure, const KIND: crate::policy::gc_work::TraceKind>(&self, __mmtk_trace: &mut T, __mmtk_objref: crate::util::ObjectReference, __mmtk_worker: &mut crate::scheduler::GCWorker<VM>) -> crate::util::ObjectReference {
+            use crate::policy::space::Space;
+            use crate::policy::gc_work::PolicyTraceObject;
+            use crate::plan::transitive_closure::PlanTraceObject;
+            #(#space_field_handler)*
+            #parent_field_delegator
+        }
+    }
+}
+
+pub fn generate_scan_object<'a>(
+    scan_object_fields: &[&'a Field],
+    ty_generics: &TypeGenerics,
+) -> TokenStream2 {
+    let scan_field_handler = scan_object_fields.iter().map(|f| {
+        let f_ident = f.ident.as_ref().unwrap();
+        let ref f_ty = f.ty;
+
+        quote! {
+            if self.#f_ident.in_space(__mmtk_objref) {
+                use crate::policy::gc_work::PolicyTraceObject;
+                <#f_ty as PolicyTraceObject #ty_generics>::scan_object::<EV>(&self.#f_ident, __mmtk_worker_tls, __mmtk_objref, __mmtk_ev);
+                return;
+            }
+        }
+    });
+
+    quote! {
+        fn scan_object<EV: crate::vm::EdgeVisitor>(&self, __mmtk_worker_tls: crate::util::opaque_pointer::VMWorkerThread, __mmtk_objref: crate::util::ObjectReference, __mmtk_ev: &mut EV) {
+            use crate::vm::Scanning;
+
+            // Plan specific
+            #(#scan_field_handler)*
+
+            // Default
+            VM::VMScanning::scan_object(__mmtk_worker_tls, __mmtk_objref, __mmtk_ev);
+        }
+    }
+}
+
+// The generated function needs to be inlined and constant folded. Otherwise, there will be a huge
+// performance penalty.
+pub fn generate_may_move_objects<'a>(
+    space_fields: &[&'a Field],
+    parent_field: &Option<&'a Field>,
+    ty_generics: &TypeGenerics,
+) -> TokenStream2 {
+    // If any space or the parent may move objects, the plan may move objects
+    let space_handlers = space_fields.iter().map(|f| {
+        let ref f_ty = f.ty;
+
+        quote! {
+            || <#f_ty as PolicyTraceObject #ty_generics>::may_move_objects::<KIND>()
+        }
+    });
+
+    let parent_handler = if let Some(p) = parent_field {
+        let ref p_ty = p.ty;
+
+        quote! {
+            || <#p_ty as PlanTraceObject #ty_generics>::may_move_objects::<KIND>()
+        }
+    } else {
+        TokenStream2::new()
+    };
+
+    quote! {
+        fn may_move_objects<const KIND: crate::policy::gc_work::TraceKind>() -> bool {
+            use crate::policy::gc_work::PolicyTraceObject;
+            use crate::plan::transitive_closure::PlanTraceObject;
+
+            false #(#space_handlers)* #parent_handler
+        }
+    }
+}
