+++ src/plan/generational/copying/global.rs
@@ -26,7 +26,7 @@ use enum_map::EnumMap;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 
-use macro_trace_object::PlanTraceObject;
+use mmtk_macro_trace_object::PlanTraceObject;
 
 #[derive(PlanTraceObject)]
 pub struct GenCopy<VM: VMBinding> {
