+++ src/plan/pageprotect/global.rs
@@ -22,7 +22,7 @@ use crate::{
 use enum_map::EnumMap;
 use std::sync::Arc;
 
-use macro_trace_object::PlanTraceObject;
+use mmtk_macro_trace_object::PlanTraceObject;
 
 #[derive(PlanTraceObject)]
 pub struct PageProtect<VM: VMBinding> {
